Updated site to include Checklist. Refactored and improved css.
Added a new button that points to the checklist app. Fixed the line underneath the link texts, it is now consistent with the width of the actual text and no hacks are required to make all links look good. Fixed a small flash of the background when loading the page.